pyproject.toml
[project]
name = "recibotou"
version = "0.1.0"
description = ""
authors = [
    {name = "biruta-san",email = "ryanhenriq2019@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn[standard] (>=0.35.0,<0.36.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "alembic (>=1.16.4,<2.0.0)",
    "pydantic[email] (>=2.11.7,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "pymysql (>=1.1.1,<2.0.0)",
    "easyocr (>=1.7.2,<2.0.0)",
    "opencv-python-headless (>=4.12.0.88,<5.0.0.0)",
    "python-dotenv (>=1.0.0,<2.0.0)",
    "gunicorn (>=22.0.0,<23.0.0)",
    "cryptography (>=46.0.3,<47.0.0)",
    "aiosmtplib (>=5.0.0,<6.0.0)",
    "google-generativeai (>=0.8.5,<0.9.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.database.dependencies]
mysqlclient = "^2.2.7"


[tool.poetry.group.dev.dependencies]
httpx = "^0.28.1"
pytest = "^8.4.1"
pytest-asyncio = "^1.1.0"

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
